Move verbose log handling into separate class
