+++ extension/socks5proxy/bin/CMakeLists.txt
@@ -4,6 +4,8 @@
 
 qt_add_executable(socksproxy 
     main.cpp
+    verboselogger.cpp
+    verboselogger.h
 )
 
 target_link_libraries(socksproxy PUBLIC
